Moved XML functions from utilities to setup to avoid cyclic dependence
