+++ pilot/user/atlas/utilities.py
@@ -10,13 +10,10 @@
 import os
 import time
 
-from xml.dom import minidom
-import xml.etree.ElementTree as ET
-
 # from pilot.info import infosys
-from pilot.user.atlas.setup import get_asetup
+from .setup import get_asetup
 from pilot.util.auxiliary import get_logger
-from pilot.util.filehandling import read_json, copy, write_file
+from pilot.util.filehandling import read_json, copy
 
 import logging
 logger = logging.getLogger(__name__)
@@ -399,93 +396,3 @@ def post_memory_monitor_action(job):
         copy(path1, path2)
     except Exception as e:
         log.warning('failed to copy memory monitor output: %s' % e)
-
-
-def create_input_file_metadata(file_dictionary, workdir, filename="PoolFileCatalog.xml"):
-    """
-    Create a Pool File Catalog for the files listed in the input dictionary.
-    The function creates properly formmatted XML (pretty printed) and writes the XML to file.
-
-    Format:
-    dictionary = {'guid': 'pfn', ..}
-    ->
-    <POOLFILECATALOG>
-    <File ID="guid">
-      <physical>
-        <pfn filetype="ROOT_All" name="surl"/>
-      </physical>
-      <logical/>
-    </File>
-    <POOLFILECATALOG>
-
-    :param file_dictionary: file dictionary.
-    :param workdir: job work directory (string).
-    :param filename: PFC file name (string).
-    :return: xml (string)
-    """
-
-    # create the file structure
-    data = ET.Element('POOLFILECATALOG')
-
-    for fileid in file_dictionary.keys():
-        _file = ET.SubElement(data, 'File')
-        _file.set('ID', fileid)
-        _physical = ET.SubElement(_file, 'physical')
-        _pfn = ET.SubElement(_physical, 'pfn')
-        _pfn.set('filetype', 'ROOT_ALL')
-        _pfn.set('name', file_dictionary.get(fileid))
-        ET.SubElement(_file, 'logical')
-
-    # create a new XML file with the results
-    xml = ET.tostring(data, encoding='utf8')
-    xml = minidom.parseString(xml).toprettyxml(indent="  ")
-    write_file(os.path.join(workdir, filename), xml)
-
-    return xml
-
-
-def get_file_info_from_xml(workdir, filename="PoolFileCatalog.xml"):
-    """
-    Return a file info dictionary based on the metadata in the given XML file.
-    The file info dictionary is used to replace the input file LFN list in the job parameters with the full PFNs
-    which are needed for direct access in production jobs.
-
-    Example of PoolFileCatalog.xml:
-
-    <?xml version="1.0" ?>
-    <POOLFILECATALOG>
-      <File ID="4ACC5018-2EA3-B441-BC11-0C0992847FD1">
-        <physical>
-          <pfn filetype="ROOT_ALL" name="root://dcgftp.usatlas.bnl.gov:1096//../AOD.11164242._001522.pool.root.1"/>
-        </physical>
-        <logical/>
-      </File>
-    </POOLFILECATALOG>
-
-    which gives the following dictionary:
-
-    {'AOD.11164242._001522.pool.root.1': ['root://dcgftp.usatlas.bnl.gov:1096//../AOD.11164242._001522.pool.root.1',
-    '4ACC5018-2EA3-B441-BC11-0C0992847FD1']}
-
-    :param workdir: directory of PoolFileCatalog.xml (string).
-    :param filename: file name (default: PoolFileCatalog.xml) (string).
-    :return: dictionary { LFN: [PFN, GUID], .. }
-    """
-
-    file_info_dictionary = {}
-    tree = ET.parse(os.path.join(workdir, filename))
-    root = tree.getroot()
-    # root.tag = POOLFILECATALOG
-
-    for child in root:
-        # child.tag = 'File', child.attrib = {'ID': '4ACC5018-2EA3-B441-BC11-0C0992847FD1'}
-        guid = child.attrib['ID']
-        for grandchild in child:
-            # grandchild.tag = 'physical', grandchild.attrib = {}
-            for greatgrandchild in grandchild:
-                # greatgrandchild.tag = 'pfn', greatgrandchild.attrib = {'filetype': 'ROOT_ALL', 'name': 'root://dcgftp.usatlas.bnl ..'}
-                pfn = greatgrandchild.attrib['name']
-                lfn = os.path.basename(pfn)
-                file_info_dictionary[lfn] = [pfn, guid]
-
-    return file_info_dictionary
