+++ pilot/util/default.cfg
@@ -68,8 +68,6 @@ looping_limit_default_user: 10800
 # The minimum allowed looping limit, 2*3600
 looping_limit_min_default: 7200
 
-# Proxy verification limit
-# ..
 # Proxy verification time (used by monitoring) in seconds
 proxy_verification_time: 600
